[auth]: * good consistent changes - tap on shoulder * works fine * TODO: extra validation on model * TODO: ERROR HANDLING

+++ src/api/items.controller.ts
@@ -1,14 +1,17 @@
 import { Request, Response, Router } from 'express';
-import { Service } from 'typedi';
+import { Inject, Service } from 'typedi';
 import authMiddleware from '../middleware/authentication';
-import { Item } from '../model/Item';
+import { IItem } from '../model/interfaces/item';
 import { ItemService } from '../service/item.service';
 
 @Service()
 export class ItemsController {
     public router: Router;
 
-    constructor(private readonly itemService: ItemService) {
+    constructor(
+        @Inject()
+        private readonly itemService: ItemService
+    ) {
         this.router = Router();
         this.initRoutes();
     }
@@ -19,25 +22,32 @@ export class ItemsController {
 
     public async listById(req: Request, res: Response) {
         const itemId: number = parseInt(req.params.id);
-        const item = await this.itemService.listById(itemId);
-        return res.send(item);
+
+        return res.send(await this.itemService.listById(itemId));
     }
 
-    public async add(req: Request, res: Response) {
-        const item: Item = await this.itemService.save(
-            new Item(req.body.title, req.body.url, req.body.imageUrl, req.body.wishlist)
+    public async save(req: Request, res: Response) {
+        const itemOptions: IItem = req.body;
+        const item: IItem = await this.itemService.save(
+            itemOptions
         );
         return res.send(item);
     }
 
     public async update(req: Request, res: Response) {
         const itemId: number = parseInt(req.params.id);
-        const item: Item = new Item(req.body.title, req.body.url, req.body.imageUrl, req.body.wishlist);
+        if(!itemId) {
+            throw new Error('An id must be provided for an update.');
+        }
+        const item: IItem = req.body;
         return res.send(await this.itemService.update(itemId, item));
     }
 
     public async delete(req: Request, res: Response) {
         const itemId: number = parseInt(req.params.id);
+        if(!itemId) {
+            throw new Error ('An id must be provided for an item deletion.');
+        }
         return res.send(this.itemService.delete(itemId));
     }
 
@@ -45,7 +55,7 @@ export class ItemsController {
     private initRoutes() {
         this.router.get('/', authMiddleware, (req, res) => this.listAll(req, res));
         this.router.get('/:id', authMiddleware, (req, res) => this.listById(req, res));
-        this.router.post('/', authMiddleware, (req, res) => this.add(req, res));
+        this.router.post('/', authMiddleware, (req, res) => this.save(req, res));
         this.router.put('/:id', authMiddleware, (req, res) => this.update(req, res));
         this.router.delete('/:id', authMiddleware, (req, res) => this.delete(req, res));
